requirements.txt
gunicorn==19.4.5
psycopg2>=2.6.1
Django==1.10.1
dj-database-url==0.4.0
pinax-theme-bootstrap==7.10.2
django-user-accounts==1.3.1
python-social-auth==0.2.21
pinax-webanalytics==2.0.4
pinax-eventlog==1.1.2
whitenoise>=3.2.2
pytz


